Converts a df into a table grob and prints turns it into a png placed in the destination

Converts a df into a table grob and prints turns it into a png placed in the destination

Usage

create_table(df, title, destination, name, dimensions)

Arguments

df

A dataframe to be converted into a table

title

A character string representing the title to place above the table

destination

A character string representing the filepath to place the png

name

A character string representing the name of the png

dimensions

A length 2 list containing the width and height of the png represented as numbers (width first)

Value

Nothing

Examples

if (FALSE) { # \dontrun{
create_table(
  data.frame(
  Age = c(25, 30, 35),
  Score = c(85.5, 90.0, 78.3)
  ),
  "TableTitle",
  "tests/testdir",
  "MyTableImage",
  list(1500,1000)
)
} # }